index-lib: identify the pagination bug - document offset and result offset are not set correctly
We are using document.id to set document offset, but that offset should be just an index into the array of all documents satisfying the query.
The document.id is also a offset, but into the 'array' of all documents in a collection.

there is no test verifying pagination, so some tests have to be written and then we should fix it